Move git integration fixtures into upper level
